requirements.txt
# openpilot Windows Viewer Dependencies
# Python 3.10+ recommended

# GUI Framework
PyQt6>=6.5.0
PyQt6-Qt6>=6.5.0

# Cap'n Proto Parser
pycapnp>=1.3.0

# DBC File Parser
cantools>=39.0.0

# Video Processing
opencv-python>=4.8.0
av>=10.0.0

# Plotting
pyqtgraph>=0.13.0

# Numerical Computing
numpy>=1.24.0
